feat(settings): 設定ページにオープンソースライセンス項目を追加
