--- feature/settings/lib/src/ui/index.dart
@@ -1,2 +1,0 @@
-export 'component/index.dart';
-export 'settings_page.dart';
