+++ feature/settings/lib/src/ui/settings/index.dart
@@ -1,0 +1,1 @@
+export 'settings_page.dart';
